Rename and apply the `AsyncThunkConfig`
