+++ frontend/src/store/thunks/auth/fetchAuthUser.ts
@@ -4,7 +4,7 @@ import { AUTH_USER_PATH } from 'config/api';
 import { User } from 'models/User';
 import { apiClient } from 'utils/api';
 import { isHttpException } from 'utils/api/errors';
-import { RejectWithValue } from '../types';
+import { AsyncThunkConfig } from '../config';
 
 export type FetchAuthUserResponse = {
   user: User;
@@ -13,7 +13,7 @@ export type FetchAuthUserResponse = {
 export const fetchAuthUser = createAsyncThunk<
   FetchAuthUserResponse,
   void,
-  { rejectValue: RejectWithValue }
+  AsyncThunkConfig
 >('auth/fetchAuthUser', async (_, thunkApi) => {
   try {
     const response = await apiClient({ intercepted: false }).get(
