+++ frontend/src/store/thunks/boards/updateTaskBoard.ts
@@ -2,7 +2,7 @@ import { createAsyncThunk } from '@reduxjs/toolkit';
 
 import { TaskBoard } from 'models';
 import { apiClient, makePath } from 'utils/api';
-import { AsyncThunkConfig } from '../types';
+import { AsyncThunkConfig } from '../config';
 
 export type UpdateTaskBoardResponse = {
   data: TaskBoard;
