+++ frontend/src/store/thunks/config/index.ts
@@ -1,2 +1,1 @@
 export * from './AsyncThunkConfig';
-export * from './RejectWithValue';
