+++ frontend/src/store/thunks/lists/createTaskList.ts
@@ -2,7 +2,7 @@ import { createAsyncThunk } from '@reduxjs/toolkit';
 
 import { TaskList } from 'models';
 import { apiClient, makePath } from 'utils/api';
-import { AsyncThunkConfig } from '../types';
+import { AsyncThunkConfig } from '../config';
 
 export type CreateTaskListResponse = {
   data: TaskList;
