--- frontend/src/store/thunks/types/RejectWithValue.ts
@@ -1,5 +1,0 @@
-export type RejectWithValue = {
-  error: {
-    message?: string;
-  };
-};
